AI-IntakeandAssessment/WorkflowsLocal/SorthaDevKit/YetToBeIntegrated(AzMigrateReportFetchfromAzure)/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
fastmcp
